backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
groq
python-dateutil
# passlib → password hashing
passlib[bcrypt]
bcrypt==4.0.1
passlib==1.7.4
# jose → JWT tokens
python-jose
reportlab